backend/requirements.txt
aiosqlite==0.16.1
argon2-cffi==20.1.0
asyncpg==0.21.0
cffi==1.14.5
click==7.1.2
flake8==3.8.4
gunicorn==20.0.4
h11==0.12.0
importlib-metadata==3.4.0
iso8601==0.1.14
mccabe==0.6.1
pycodestyle==2.6.0
pycparser==2.20
pydantic==1.7.3
pyflakes==2.2.0
PyPika==0.44.1
pytz==2020.5
six==1.15.0
starlette==0.14.2
tortoise-orm==0.16.21
typing-extensions==3.7.4.3
uvicorn==0.13.3
zipp==3.4.0
